refactor: search transactions queries
